sample-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.38
asyncpg==0.30.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.6
python-dotenv==1.0.1
azure-identity==1.15.0
azure-storage-blob==12.19.0
opencensus-ext-azure==1.1.13
opencensus>=0.11.4,<1.0.0
httpx==0.26.0
